+++ ingenialink/servo.py
@@ -9,7 +9,7 @@ import ingenialogger
 import numpy as np
 
 from ingenialink.bitfield import BitField
-from ingenialink.canopen.dictionary import CanopenDictionaryV2
+from ingenialink.canopen.dictionary import CanopenDictionaryV2, CanopenDictionaryV3
 from ingenialink.configuration_file import ConfigRegister, ConfigurationFile
 from ingenialink.constants import (
     DEFAULT_DRIVE_NAME,
@@ -31,8 +31,12 @@ from ingenialink.dictionary import (
 from ingenialink.emcy import EmergencyMessage
 from ingenialink.enums.register import RegAccess, RegAddressType, RegDtype
 from ingenialink.enums.servo import ServoState
-from ingenialink.ethercat.dictionary import EthercatDictionaryV2
-from ingenialink.ethernet.dictionary import EthernetDictionaryV2
+from ingenialink.ethercat.dictionary import EthercatDictionaryV2, EthercatDictionaryV3
+from ingenialink.ethernet.dictionary import (
+    EoEDictionaryV3,
+    EthernetDictionaryV2,
+    EthernetDictionaryV3,
+)
 from ingenialink.exceptions import (
     ILAccessError,
     ILConfigurationError,
@@ -84,7 +88,14 @@ class DictionaryFactory:
         """
         major_version, _ = cls.__get_dictionary_version(dictionary_path)
         if major_version == 3:
-            return DictionaryV3(dictionary_path, interface)
+            if interface == Interface.CAN:
+                return CanopenDictionaryV3(dictionary_path)
+            if interface == Interface.ECAT:
+                return EthercatDictionaryV3(dictionary_path)
+            if interface == Interface.EoE:
+                return EoEDictionaryV3(dictionary_path)
+            if interface in [Interface.ETH, Interface.VIRTUAL]:
+                return EthernetDictionaryV3(dictionary_path)
         if major_version == 2:
             if interface == Interface.CAN:
                 return CanopenDictionaryV2(dictionary_path)
@@ -94,7 +105,9 @@ class DictionaryFactory:
                 return EthernetDictionaryV2(dictionary_path)
             if interface == Interface.VIRTUAL:
                 return VirtualDictionary(dictionary_path)
-        raise NotImplementedError(f"Dictionary version {major_version} is not supported")
+        raise NotImplementedError(
+            f"Dictionary version {major_version} is not supported for interface {interface.name}"
+        )
 
     @classmethod
     def get_dictionary_description(
@@ -1263,7 +1276,6 @@ class Servo:
         reg: Union[str, Register],
         data: Union[int, float, str, bytes],
         subnode: int = 1,
-        **kwargs: Any,
     ) -> None:
         """Writes a data to a target register.
 
@@ -1271,7 +1283,6 @@ class Servo:
             reg: Target register to be written.
             data: Data to be written.
             subnode: Target axis of the drive.
-            **kwargs: Keyword arguments.
 
         Raises:
             ILAccessError: Wrong access to the register.
@@ -1281,18 +1292,19 @@ class Servo:
         if _reg.access == RegAccess.RO:
             raise ILAccessError("Register is Read-only")
         data_bytes = data if isinstance(data, bytes) else convert_dtype_to_bytes(data, _reg.dtype)
-        self._write_raw(_reg, data_bytes, **kwargs)
+        self._write_raw(_reg, data_bytes)
         self._notify_register_update(_reg, data)
 
     def read(
-        self, reg: Union[str, Register], subnode: int = 1, **kwargs: Any
+        self,
+        reg: Union[str, Register],
+        subnode: int = 1,
     ) -> Union[int, float, str, bytes]:
         """Read a register value from servo.
 
         Args:
             reg: Register.
             subnode: Target axis of the drive.
-            **kwargs: Keyword arguments.
 
         Returns:
             int, float or Value stored in the register.
@@ -1305,10 +1317,40 @@ class Servo:
         if access == RegAccess.WO:
             raise ILAccessError("Register is Write-only")
 
-        raw_read = self._read_raw(_reg, **kwargs)
+        raw_read = self._read_raw(_reg)
+
         value = convert_bytes_to_dtype(raw_read, _reg.dtype)
         self._notify_register_update(_reg, value)
         return value
+
+    def write_complete_access(
+        self, reg: Union[str, Register], data: bytes, subnode: int = 1
+    ) -> None:
+        """Write a complete access register.
+
+        Args:
+            reg: Register to be written.
+            data: Data to be written.
+            subnode: Target subnode of the drive.
+        """
+        _reg = self._get_reg(reg, subnode)
+        self._write_raw(_reg, data, complete_access=True)
+
+    def read_complete_access(
+        self, reg: Union[str, Register], subnode: int = 1, buffer_size: int = 0
+    ) -> bytes:
+        """Read a complete access register.
+
+        Args:
+            reg: Register to be read.
+            subnode: Target subnode of the drive.
+            buffer_size: Size of the buffer to read.
+
+        Returns:
+            Data read from the register.
+        """
+        _reg = self._get_reg(reg, subnode)
+        return self._read_raw(_reg, buffer_size=buffer_size, complete_access=True)
 
     def read_bitfields(
         self,
@@ -1509,12 +1551,13 @@ class Servo:
         return self.write(self.DIST_DATA, subnode=0, data=data)
 
     @abstractmethod
-    def _write_raw(self, reg: Register, data: bytes) -> None:
+    def _write_raw(self, reg: Register, data: bytes, **kwargs: Any) -> None:
         """Write raw bytes to a target register.
 
         Args:
             reg: Target register to be written.
             data: Data to be written.
+            **kwargs: Additional arguments for the write operation.
 
         Raises:
             ILIOError: Error writing the register.
@@ -1541,11 +1584,12 @@ class Servo:
         raise NotImplementedError
 
     @abstractmethod
-    def _read_raw(self, reg: Register) -> bytes:
+    def _read_raw(self, reg: Register, **kwargs: Any) -> bytes:
         """Read raw bytes from a target register.
 
         Args:
             reg: Register.
+            kwargs: Additional arguments for the read operation.
 
         Returns:
             Raw bytes reading from servo.