+++ tests/canopen/test_canopen_dictionary_v3.py
@@ -4,7 +4,8 @@ import pytest
 
 from ingenialink import CanopenRegister
 from ingenialink.bitfield import BitField
-from ingenialink.dictionary import CanOpenObjectType, DictionaryV3, Interface, SubnodeType
+from ingenialink.canopen.dictionary import CanopenDictionaryV3
+from ingenialink.dictionary import CanOpenObjectType, Interface, SubnodeType
 from ingenialink.exceptions import ILDictionaryParseError
 
 path_resources = "./tests/resources/canopen/"
@@ -29,7 +30,7 @@ def test_read_dictionary():
         "image": None,
     }
 
-    canopen_dict = DictionaryV3(dictionary_path, Interface.CAN)
+    canopen_dict = CanopenDictionaryV3(dictionary_path)
 
     for attr, value in expected_device_attr.items():
         assert getattr(canopen_dict, attr) == value
@@ -40,7 +41,7 @@ def test_read_dictionary_file_not_found():
     dictionary_path = "false.xdf"
 
     with pytest.raises(FileNotFoundError):
-        DictionaryV3(dictionary_path, Interface.CAN)
+        CanopenDictionaryV3(dictionary_path)
 
 
 @pytest.mark.no_connection
@@ -60,7 +61,7 @@ def test_read_dictionary_registers():
         1: ["COMMU_ANGLE_SENSOR", "DRV_STATE_CONTROL"],
     }
 
-    canopen_dict = DictionaryV3(dictionary_path, Interface.CAN)
+    canopen_dict = CanopenDictionaryV3(dictionary_path)
 
     for subnode in expected_regs_per_subnode:
         subnode_registers = canopen_dict.registers(subnode)
@@ -74,7 +75,7 @@ def test_read_dictionary_registers_multiaxis():
     expected_num_registers_per_subnode = {0: 4, 1: 1, 2: 1}
     dictionary_path = join_path(path_resources, dict_can_v3_axis)
 
-    canopen_dict = DictionaryV3(dictionary_path, Interface.CAN)
+    canopen_dict = CanopenDictionaryV3(dictionary_path)
     assert canopen_dict.subnodes == {
         0: SubnodeType.COMMUNICATION,
         1: SubnodeType.MOTION,
@@ -93,7 +94,7 @@ def test_read_dictionary_categories():
     ]
     dictionary_path = join_path(path_resources, dict_can_v3)
 
-    canopen_dict = DictionaryV3(dictionary_path, Interface.CAN)
+    canopen_dict = CanopenDictionaryV3(dictionary_path)
 
     assert canopen_dict.categories.category_ids == expected_categories
 
@@ -106,7 +107,7 @@ def test_read_dictionary_errors():
     ]
     dictionary_path = join_path(path_resources, dict_can_v3)
 
-    canopen_dict = DictionaryV3(dictionary_path, Interface.CAN)
+    canopen_dict = CanopenDictionaryV3(dictionary_path)
 
     assert list(canopen_dict.errors) == expected_errors
 
@@ -119,7 +120,7 @@ def test_read_xdf_register():
     reg_id = "DRV_DIAG_ERROR_LAST_COM"
     subnode = 0
 
-    canopen_dict = DictionaryV3(dictionary_path, Interface.CAN)
+    canopen_dict = CanopenDictionaryV3(dictionary_path)
     target_register = canopen_dict.registers(subnode)[reg_id]
 
     assert isinstance(target_register, CanopenRegister)
@@ -130,7 +131,7 @@ def test_read_xdf_register():
 @pytest.mark.no_connection
 def test_object():
     dictionary_path = join_path(path_resources, dict_can_v3)
-    canopen_dict = DictionaryV3(dictionary_path, Interface.CAN)
+    canopen_dict = CanopenDictionaryV3(dictionary_path)
     canopen_object = canopen_dict.get_object("CIA301_COMMS_RPDO1_MAP", 0)
     assert canopen_object.uid == "CIA301_COMMS_RPDO1_MAP"
     assert canopen_object.object_type == CanOpenObjectType.RECORD
@@ -147,7 +148,7 @@ def test_object():
 @pytest.mark.no_connection
 def test_object_not_exist():
     dictionary_path = join_path(path_resources, dict_can_v3)
-    canopen_dict = DictionaryV3(dictionary_path, Interface.CAN)
+    canopen_dict = CanopenDictionaryV3(dictionary_path)
     with pytest.raises(KeyError):
         canopen_dict.get_object("NOT_EXISTING_UID", 0)
 
@@ -155,7 +156,7 @@ def test_object_not_exist():
 @pytest.mark.no_connection
 def test_safety_pdo_not_implemented():
     dictionary_path = join_path(path_resources, dict_can_v3)
-    canopen_dict = DictionaryV3(dictionary_path, Interface.CAN)
+    canopen_dict = CanopenDictionaryV3(dictionary_path)
     with pytest.raises(NotImplementedError):
         canopen_dict.get_safety_rpdo("NOT_EXISTING_UID")
     with pytest.raises(NotImplementedError):
@@ -167,7 +168,7 @@ def test_wrong_dictionary():
     with pytest.raises(
         ILDictionaryParseError, match="Dictionary cannot be used for the chosen communication"
     ):
-        DictionaryV3("./tests/resources/test_dict_ecat_eoe_v3.0.xdf", Interface.CAN)
+        CanopenDictionaryV3("./tests/resources/test_dict_ecat_eoe_v3.0.xdf", Interface.CAN)
 
 
 @pytest.mark.no_connection
@@ -190,7 +191,7 @@ def test_register_default_values(dictionary_path):
             "COMMU_ANGLE_SENSOR": 4,
         },
     }
-    canopen_dict = DictionaryV3(dictionary_path, Interface.CAN)
+    canopen_dict = CanopenDictionaryV3(dictionary_path)
     for subnode, registers in canopen_dict._registers.items():
         for register in registers.values():
             assert register.default == expected_defaults_per_subnode[subnode][register.identifier]
@@ -220,7 +221,7 @@ def test_register_description(dictionary_path):
             "COMMU_ANGLE_SENSOR": "Indicates the sensor used for angle readings",
         },
     }
-    canopen_dict = DictionaryV3(dictionary_path, Interface.CAN)
+    canopen_dict = CanopenDictionaryV3(dictionary_path)
     for subnode, registers in canopen_dict._registers.items():
         for register in registers.values():
             assert (
@@ -232,7 +233,7 @@ def test_register_description(dictionary_path):
 @pytest.mark.no_connection
 def test_register_bitfields():
     dictionary_path = join_path(path_resources, dict_can_v3)
-    canopen_dict = DictionaryV3(dictionary_path, Interface.CAN)
+    canopen_dict = CanopenDictionaryV3(dictionary_path)
 
     for registers in canopen_dict._registers.values():
         for register in registers.values():
@@ -253,7 +254,7 @@ def test_register_bitfields():
 @pytest.mark.no_connection
 def test_register_is_node_id_dependent():
     dictionary_path = join_path(path_resources, dict_can_v3)
-    canopen_dict = DictionaryV3(dictionary_path, Interface.CAN)
+    canopen_dict = CanopenDictionaryV3(dictionary_path)
     assert canopen_dict.registers(0)["CIA301_COMMS_RPDO1_1"].is_node_id_dependent
     assert not canopen_dict.registers(0)["CIA301_COMMS_RPDO1_2"].is_node_id_dependent
     assert not canopen_dict.registers(0)["CIA301_COMMS_RPDO1_3"].is_node_id_dependent